day0331

+++ src/main/java/com/dqv5/backstage/dao/SysLogDao.java
@@ -9,7 +9,7 @@ import java.util.List;
  */
 public interface SysLogDao {
 
-    SysLog getByPrimaryKey(Long id);
+    SysLog getByPrimaryKey(Integer id);
 
     void save(SysLog record);
